simple diamond
Create the C program q1a_simple_diamond.c to produce a program which takes 1 argument, the height H of the diamond. Prints a diamond which is made up of H rows of asterisk (a.k.a. star, *) characters and spaces following this specification:
• The diamond’s top and bottom rows must both be a single asterisk (same line if H=1)
• Subsequent rows must grow or shrink by 2 asterisks, according to the overall shape
• A single row in the middle of the diamond must have exactly H asterisks (no spaces in this one)
• The first asterisk in the middle row must appear at the very beginning of the line
• The asterisks in all rows must be perfectly centered with respect to the middle row
Your program is never allowed to segmentation fault, no matter how it is run. To prevent this, input checking must be performed:
• The program needs exactly one argument after the program name, no more and no less. Otherwise print: “ERROR: Wrong number of arguments. One required.”
• The height argument must be a positive integer and odd (ensures the overall shape works). Otherwise print: “ERROR: Bad argument. Height must be positive odd integer.”